after_client_testing

+++ database/migrations/2019_01_31_091857_create_sales_table.php
@@ -20,8 +20,8 @@ class CreateSalesTable extends Migration
             $table->unsignedInteger('saletype_id');
             // $table->foreign('employee_id')->references('id')->on('employees');
             $table->unsignedInteger('jobnumber')->unique();;
-            $table->unsignedInteger('amount');
-            $table->unsignedInteger('commission');
+            $table->float('amount');
+            $table->float('commission');
             $table->date('dateofsale');
             $table->softDeletes();
             $table->timestamps();
