+++ database/migrations/2019_02_01_164355_create_commissions_table.php
@@ -15,7 +15,7 @@ class CreateCommissionsTable extends Migration
     {
         Schema::create('commissions', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('commission')->unsigned();
+            $table->float('commission')->unsigned();
             $table->timestamps();
         });
     }
